requirements.txt
openai==1.55.3
PyYAML==6.0.2